+++ ci/licenses_golden/licenses_flutter
@@ -916,6 +916,8 @@ ORIGIN: ../../../flutter/fml/message_loop_task_queues_benchmark.cc + ../../../fl
 ORIGIN: ../../../flutter/fml/native_library.h + ../../../flutter/LICENSE
 ORIGIN: ../../../flutter/fml/paths.cc + ../../../flutter/LICENSE
 ORIGIN: ../../../flutter/fml/paths.h + ../../../flutter/LICENSE
+ORIGIN: ../../../flutter/fml/platform/android/cpu_affinity.cc + ../../../flutter/LICENSE
+ORIGIN: ../../../flutter/fml/platform/android/cpu_affinity.h + ../../../flutter/LICENSE
 ORIGIN: ../../../flutter/fml/platform/android/jni_util.cc + ../../../flutter/LICENSE
 ORIGIN: ../../../flutter/fml/platform/android/jni_util.h + ../../../flutter/LICENSE
 ORIGIN: ../../../flutter/fml/platform/android/jni_weak_ref.cc + ../../../flutter/LICENSE
@@ -3660,6 +3662,8 @@ FILE: ../../../flutter/fml/message_loop_task_queues_benchmark.cc
 FILE: ../../../flutter/fml/native_library.h
 FILE: ../../../flutter/fml/paths.cc
 FILE: ../../../flutter/fml/paths.h
+FILE: ../../../flutter/fml/platform/android/cpu_affinity.cc
+FILE: ../../../flutter/fml/platform/android/cpu_affinity.h
 FILE: ../../../flutter/fml/platform/android/jni_util.cc
 FILE: ../../../flutter/fml/platform/android/jni_util.h
 FILE: ../../../flutter/fml/platform/android/jni_weak_ref.cc
