--- fml/platform/android/cpu_affinity_unittests.cc
@@ -1,18 +1,0 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-#include "cpu_affinity.h"
-
-#include "gtest/gtest.h"
-#include "logging.h"
-
-namespace fml {
-namespace testing {
-
-TEST(CpuAffinity, CanDoThing) {
-  EXPECT_EQ(true, false);
-}
-
-}  // namespace testing
-}  // namespace fml
